Fix symlink handling in doCopyFileOrDir()
Symlinks need special treatment when we're copying them into the
report. The target path stored in the report must always be
relative but we need to pass an absolute path for the target to
the recursive doCopyFileOrDir() call that picks up the target for
us.

There are further problems with the current code but these cannot
be fixed trivially: symbolic links to directories are currently
ignored but are fundamental to the layout of file systems like sys
(and to a lesser extent proc). Supporting this properly requires
an algorithm that can cope with arbitrary symlink loops in the
tree being copied and that correctly copies-in any links that may
point outside of the tree currently being copied.

See Issues #71, #72 for more details.