Fix parameter handling and pg_dump collection in psql
Re-work the parameter usage and pg_dump collection in the
PostgreSQL plug-in so that it works correctly on systems without
postgres even when the user specifies -o psql with or without -a.

The pgdump tarballs are no longer created in tmp and collected
using addCopySpec(); it's better to just send them straight into
the sos_commands path in the first place (avoids the need for a
postproc() method to do cleanup).

Resolves: bz852049