Improve policy template handling
The current boilerplate text in the policy class has inconsistent
formatting and does not make sense with some distribution names
(e.g. Red Hat Enterprise Linux).

Fix this by storing the string in paragraphs and formatting it
with textwrap and adding new tags to make the construction of the
message more flexible:

  vendor
  vendor_url
  vendor_text
  tmpdir

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>