Rename --silent to --quiet
'silent' is a poort name for an option that admits to not being
silent:

  --silent              Only display FATAL errors on stdout

Rename the option to '--quiet' instead and document it in the man
page. It's probably also worth either documenting the fact that
UI text is still printed when --quiet is given in non-batch mode
or changing interactive mode to only print essential prompts when
run with --quiet.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>