Update SELinux plug-in for modern distributions
The old SELinux module has had no maintenance in some time. Add
the output of the following commands:

 - sestatus -b [ SELinux status and booleans ]
 - selinuxdefcon root
 - selinuxconlist root
 - selinuxexeccon /bin/passwd

Add '-v' to fixfiles command line to produce more verbose output.

Drop all setroubleshoot support - this is an optional package for
all supported distributions and should be handled in a separate
sos plug-in.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>