Move command line parsing out of SoSReport class
Separate the command line logic from the rest of the SoSReport
class and add a new SoSOptions class to contain it.

SoSOptions may either be initialised from a command line by
passing in an argument list to the initializer or by creating
an empty object and setting values for individual properties.

For now a SoSOptions object is initialized in the SoSReport
initializer (from the argument list passed from __main__).
Eventually this will move into __main__ allowing SoSOptions and
SoSReport to become an interface for external users.

The next step is to move all input validation into the SoSOptions
class (it's currently spread around the code that uses the
options).

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>