Fix detection of ext file systems in filesys plug-in
The code to collect dumpe2fs output does not work on modern
systems due to changes in the output of the mount command.

Have the plug-in use /proc/mounts instead and simplify and improve
the regex used to find devices containing ext[234] file systems
and switch to 'dumpe2fs -h' collection rather than getting the
full group lists.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>