Move all block device data from filesys to block module
The filesys module has long had a lot of (very weird) code for
detecting and processing a list of whole-disk devices in the
system. Apart from being garbage this doesn't really belong in
the file systems module - move it to block and replace the
hokey /proc-and-hdparm-or-regex thing with the existing directory
walk over /sys/block.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>