Move dmesg collection from general to kernel
The kernel ring buffer is certainly more 'kernel' than 'general'.
Put it where it belongs and remove the duplicate dmesg call and
'dmesg_now' suggested filename (there's no need as it's in a
separate directory to the saved version at var/log/dmesg and the
normal action of dmesg is to print the current content).

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>