Fix latest debian packaging suggestions.
1) lintian information, you should run lintian with '-IE --pedantic',
lintian will tell you more if there are any problem. There are still
some lintian information, but I don't care of them. you may fix them
when you polish your package.

2) You use old format copyright files, it is better to switch to
DEP5[1] compatible format. lintian shows this information.

3) The section is python, but I think admin should be better, after
all, sosreport is a sysadmin tool

4) Please move LICENSE file to usr/share/doc/sosreport, rather than
remove it.

A note about 4. I have removed the actual installing of the LICENSE file
in the Makefile in favor of the distribution specific packaging handle it.

For example, within Debian everything is placed in debian/copyright which
removes the need for our Makefile to do anything with the LICENSE file.

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>