Always use '-f-' when calling tar
The tar program evaluates TAPE to determine the default tape
device. This means that when sos is run with this environment
variable set it may write the report data to the tape device
and fail to generate a correct archive.

This may cause a loss of data from the tape device as well as
a loss of diagnostic data.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>