Fix Plugin.add_copy_spec_limit() checks
The logic in add_copy_spec_limit() only attempted to collect a
'tailed' version of a file that is over the size limit in the case
that it is the first file in the list of specs.

Change this to always collect the tailed version for the first
file to exceed the size limit and improve the variable naming
to be more consistent with related methods.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>