Fix broken binary checks in openstack.py
The OpenStack plug-in attempts to check for the presence of
binaries before running them. To do this it passes a relative
path to os.path.exists(); clearly this will never succeed.

Ideally plug-ins should never need to do these checks. If the
binary exists we'll run it and if not we won't (with no harmful
side-effects). This is how sos is currently designed and if there
is a need for plug-ins to ever do this we should add a method to
the base plug-in class e.g. Plugin.find_command() to do this
properly (i.e. evaluating the policy-supplied PATH).

For now, just make all of this unconditional.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>