Restrict wbinfo collection to the current domain.
The samba module calls the wbinfo command to collect user and
group information. This by default will search all domains to
which the host has access. In a large AD environment with a number
of trusted domains this causes the winbindd process to consume
excessive CPU and memory resources.

The Samba developers have commented that fixing this would require
a complete rewrite of the winbind protocol and wbinfo client.

Since listing out the _entire_ set of visible users is unlikely
to be of any use anyway pass the --domain='.' option to restrict
the search to the current domain of the system.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>