Fix traceback when sar module is force-enabled
The sar plug-in assumes that when it runs the /var/lib/sa
directory and binaries exist (they are part of checkenabled() for
this module).

This leads to a traceback when the plug-in lists a non-existent
directory on hosts where sar is not installed but '-o sar' is
specified on the command line.

Catch the exception and exit gracefully.

(Edited by bmr to improve error message logged).

Signed-off-by: David Kutálek <dkutalek@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>