Force LC_ALL=C for external commands
Backport of b3d4b9f.

Commands run by sos inherit the environment of the user running the
program. This includes locale settings meaning that collected
output is subject to local language, sorting and formatting
customisations that can be undesirable when post-processing or
parsing of the gathered data is required.

Force all external processes to use the 'C' locale.

Additionally on rhel-6 branch remove the LANG=C hack from the sar
plug-in.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>