Make 'ipsec barf' collection non-default in openswan plug-in
Backport of 4858ce2.

The 'ipsec barf' debugging command has a nasty bug where it runs
grep on /var/log/lastlog. Since this can be a multi-gigabyte
sparse file this doesn't work very well and consumes huge amounts
of memory. This can lead to timeouts and aborts when running sos
since the openswan plug-in calls this command by default.

The barf information is seldom useful anyway - turn it off by
default to avoid this problem.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>