Add restricted rpm verify
Backport of commit 1b3fc3f.

Add a default call to rpm -V/--verify for a targeted subset of
packages. This greatly reduces the time taken relative to rpm -Va
while still capturing important diagnostic information.

In future this capability could be made a feature of the
PackageManager class interface and available to all ports/modules.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>