Add Plugin method to build command output paths
Add a convenience method to construct an absolute path for
command output from a relative path or file name. This can be
used for e.g. to pass an output path to an external program that
produces output at a named path rather than on stdout and can be
used internally by the Plugin.make_cmd_dirs() method.

Convert the SELinux plug-in to use the new method when passing
output paths to the semanage command.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>