Python 3 port
This includes a necessary dependency on python-six for its compability layer
since we are wanting to continue support for both Python 2.7.x and Python 3.x.
In addition, this will allow us to effectively phase out Python 2 support
when/if the time arises that all interested distributions have done away with
Python 2.

This port passes all unittests for both python 2.7.x and python 3.3.x

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>