Restore --build command line option
Releases of sos prior to 3.0 included the '--build' option to
disable creation of a compressed archive and to leave the
temporary build tree in place. This was removed with the
reorganisation of the archive classes to support in-line
archiving. Since all supported policies are now using an archive
that derives from FileCacheArchive (commit 0178d5f) the option
can be re-introduced.

Archive classes that do not accumulate files in a temporary
directory will return the path to the in-progress archive file.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>