Make ethernet interface detection more robust
The networking module parses the link information in the output
of the 'ip -o addr' command to determine a list of ethernet
interfaces on which to run ethtool.

An upstream change in the iproute package appears to have
introduced a regression in the output of this command which causes
link information to not be displayed:

2: eth0: <BROADCAST,MULTICAST,UP,LOWER_UP> mtu 1500 qdisc pfifo_fast state UP qlen 1000\    link/ether 52:54:00:ce:9d:6d brd ff:ff:ff:ff:ff:ff
2: eth0    inet 192.168.122.31/24 brd 192.168.122.255 scope global eth0\       valid_lft forever preferred_lft forever
2: eth0    inet6 fe80::5054:ff:fece:9d6d/64 scope link \       valid_lft forever preferred_lft forever

vs:

2: eth0    inet 192.168.122.31/24 brd 192.168.122.255 scope global eth0\       valid_lft forever preferred_lft forever
2: eth0    inet6 fe80::5054:ff:fece:9d6d/64 scope link \       valid_lft forever preferred_lft forever

This breaks detection of ethernet interfaces and all ethtool data
collection.

This was introduced in the following iproute commit:

    commit af9d406f99853848054162ed7aefbe71dc03e433
    Author: Mike Frysinger <vapier@gentoo.org>
    Date:   Mon Aug 13 08:09:52 2012 -0700

        Fix regression with 'ip address show'

Which appears to fix a different problem (i.e. the change to link
info was unintended).

Make this more robust by switching the networking module to use
'ip -o link' instead which explicitly requests the required
information.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>