Tidy up formatting of UI preamble text
The formatting of the template preamble used a narrow terminal
width since there is no native python method for obtaining the
actual terminal geometry and the common methods for doing this
are non-portable (e.g. Windows specific ioctls or shelling out
to stty).

This causes the top of the message to scroll off the screen on
a standard 80x25 terminal (e.g. putty or a non-resized
gnome-terminal or xterm window).

Remove the paragraph indentation and increase the assumed width
from 58 to 72 chars. This causes the message to fill a standard
80x25 terminal without scrolling.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>