Fix cluster module crm_report support
The cluster plugin used an obsolete sos-2.2 method to determine
the command output directory. This causes an excaption at runtime
since the referenced properties no longer exist.

The crm_report script also expects a --from date and will not
collect data unless this is passed. Default to passing a value 72
hours before the current time and add a 'crm_from' option to the
cluster module to allow the user to override this.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>