Always treat rhevm vdsmlogs option as string
The rhevm plugin has a 'vdsmlogs' option to pass in a set of log
files to be collected. When run with '-a' (or with
'-krhevm.vdsmlogs) this evaluates to 'True' (a bool) and causes
an exception when this is passed to add_copy_specs().

Always treat the value as a string.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>