Make get_cmd_output_now() behaviour match 2.2
The equivalent method in sos-2.2, collectOutputNow() returned an
absolute path to the collected file. Since the archive changes in
3.0 this now returns a path relative to the root of the archive.

This breaks existing users of the interface that try to open and
process the content of the collected file (e.g. gluster).

Return a join of the archive path and file path to the caller.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>