Collect network interfaces flags
The output of ifconfig or ip link is not enough to understand the full
state of the network card flags. In particular IFF_PROMISC is set on a
a bridge port-member but that is not displayed via ifconfig nor ip link.

See also https://bugzilla.redhat.com/show_bug.cgi?id=199979c#17 why we
need to do this.