Clean up sos_get_command_output() and friends
Returning a 3-tuple has always been ugly. It gets worse as the
parameter list for this family of functions grows. Worse, the 3rd
member of the tuple is unused and is always set to 0.

Rip the whole mess out and return a single, simple dictionary
object with 'status' and 'output' keys.

Update utilities_tests to reflect the new interface.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>