Record duration of calls to external programs at debug level
The old profiling report was removed in favour of using external
profilers like cProfile. This is useful for overall profiling and
identifying code hotspots but is less helpful for drilling down
to the collection of individual items.

Since running programs is the dominant cost on typical hosts log
the time taken to run each call and tag it with the binary name.

This allows looking at the per-call and per-command cost across a
run of sosreport.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>