[plugintools] preserve permissions on all path components
When a path is added to the archive via addCopySpec() we copy the
content itself with shutil.copy2() which propagates permissions
from the source to the destination.

This ignores any directory components in the path leading up to
the item to be copied. For e.g.:

  self.addCopySpec("/foo/bar/baz/qux")

Will collect the file 'qux' and preserve its permissions vai a
shutil.copy2() call but '/foo/bar/baz' will be created via an
os.makedirs() call (receiving whatever permissions the current
umask allows).

Fix this by replacing the call to os.makedirs() with a call to
Plugin.__copydirs() which will walk the path from root down to the
final containing directory calling makedirs() and copystat() as
appropriate to copy permissions for the entire path.