[rpm] reduce verify timeout to 180s and -Va timeout to 600s
The initial rpm verify timeout was set to 3600s (one hour). This
is insane; if a verify has not completed in that time then it's
unlikely that it ever will. With the current set of verify
patterns the total time taken in rpm (including both the verify
and all other runs, e.g. policy initialisation) is under 60s on
all my test hosts (bare metal and virtual).

Allow a factor of three over this figure and set the timeout to
180s.