[plugins] improve plugin class docstrings
Plugin class docstrings should be a brief (42 characters or less)
description of the component the plugin supports. This is needed
since the class docstring is displayed in the table of enabled
plugins when running 'sosreport --list-plugins'.

Class docstrings should not describe the distribution a particular
class runs on; this is redundant since only plugins tagges with a
class matching the running policy will be displayed.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>