[sosreport] initial profiles support
Add simple profile support. Plugins can define a 'profiles'
attribute that contains a list of profile names that this plugin
belongs to. If a profile is given on the command line only plugins
belonging to that profile will run.

Two new command-line options are introduced in thie commit:

  --list-profiles
  --profile=NAME

Profiles can be combined with -l/--list-plugins in order to see
the effect of the profile on the set of enabled plugins.

Fixes #247.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>