[sosreport] fix silent exception handling in sosreport
The OSError and IOError exception branches for the setup, collect,
and postproc methods silently ignore plugin exceptions that are
not on the list of fatal fs errors (that will cause sos to halt
immediately).

Fix this and either raise the exception (if --debug is given) or
log it to a file via _log_plugin_exception().

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>