[mysql] test for boolean values in user and password options
If sosreport is run with '-a' all options will be set to boolean
True. This causes an exception if an attempt is made to set an
environment variable to the option value:

  Traceback (most recent call last):
    File "/usr/sbin/sosreport", line 25, in <module>
      main(sys.argv[1:])
    File "/usr/lib/python2.7/site-packages/sos/sosreport.py", line 1435, in main
      sos.execute()
  TypeError: must be string, not bool

  > /usr/lib64/python2.7/os.py(471)__setitem__()
  -> putenv(key, item)

Test both values with isinstance(val, bool) and do not attempt to
collect a database dump if either is a boolean.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>