[plugin] handle ELOOP in _copy_dir()
A problem with systemd's management of the binfmt_misc automount
point in Atomic environments causes attempts to access this path to
fail with ELOOP:

  >>> import os
  >>> os.listdir("/host/proc/sys/fs/binfmt_misc/")
  Traceback (most recent call last):
    File "<stdin>", line 1, in <module>
  OSError: [Errno 2] No such file or directory: '/host/proc/sys/fs/binfmt_misc/'

For reasons that are not yet clear this causes the entire sos
process to immediately terminate.

For now avoid the problem by enclosing the problem os.listdir in
a try/except block that explicitly handles the one errno value
implicated here.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>