[foreman] Skip collection of generic resources
New option was added to foreman-debug (-g) to skip collecting of generic
resources (firewall, selinux, cpu, memory). This patch leverages that.

https://github.com/theforeman/foreman/blob/develop/script/foreman-debug

Since katello-debug is now part of foreman-debug (as a plugin), there is no
need of separate sosreport plugin.

https://github.com/Katello/katello/blob/master/deploy/script/katello-debug.sh

New subpackage called foreman-debug was introduced. It is present in both
upstream Foreman and Katello installations as well as on both Satellite 6 main
node and capsule instances. It can be used to detect if foreman-debug script is
present.

Signed-off-by: Lukas Zapletal <lzap+git@redhat.com>