[cluster] remove some files and commands that are no longer needed
Removed some of the files and commands that are no longer needed on
RHEL6+.  In addition, the gfs_lockdump option did not capture gfs2 lock
dumps.  The option is now called gfs2_lockdump and will mount the
/sys/kernel/debug directory and the gfs2 lockdumps will be copied from
that mount point.

In addition added a couple files that are needed for pacemaker/dlm.

Resolves: rhbz#1083656

Signed-off-by: Shane Bradley <sbradley@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>