[juju] Update log collection to have limits
Juju can name files differently depending on the deployment
so we need iterate over all the files and capture at least some
of them all. At the same time the files can get quite big so we
need to be able to limit their size.

Now captures upstart/juju-db.log which is usually quite small.
Juju local is usually used for testing so we just capture
all-machines.log for that (format /var/log/juju-name-name/)

Lastly we capture the ls of all of the key directories just in
case we missed something.

Signed-off-by: Bryan Quigley <bryan.quigley@canonical.com>