[openstack] New openstack_instack plugin.
Add a new plugin for instack [1] and instack-undercloud [2].
These two components are currently used to drive the deployment of
the undercloud in RDO Manager and RHEL OSP director environments.

Instack itself is considered low level tooling and is in turn called
by higher level user applications such as the RDO Manager Openstack
client plugin [3].

This plugin also adds a new openstack_undercloud profile to be used
to tag simillar Openstack plugins that are only found in the undercloud.

[1] https://github.com/rdo-management/instack
[2] https://github.com/rdo-management/instack-undercloud
[3] https://github.com/rdo-management/python-rdomanager-oscplugin

Resolves #620.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>