[archive] Force decoding if content is bytes
When content is coming from the output of the tail()
method, it is a bytes in python3 and a string in python2.

Force the bytes value to be decoded otherwise the later
write() function will fail and nothing will be written to
the file.

Conditional decoding is necessary as the strings in python3
do not have a decode method.
Closes #586

Signed-off-by: Louis Bouchard <louis.bouchard@canonical.com>