[sosreport] move archive checksumming to sosreport
Although the digest algorithm is policy controlled the actual
mechanism to checksum the archive does not belong in the policies
module: historically this was done to keep the code that calculates
the checksum close to the UI code that reports it.

Move the calculation to the main SoSReport class's final_work()
method and add a 'checksum' argument to the display_results()
method so that the value can be reported.

In future it may make sense to push the checksum code directly into
the archive class.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>