[docker] Gather more data and expand plugin options
This patch changes the behavior of the docker plugin and collects a bit more
data than before. In addition, it also adds 'docker-engine' to the package lists
as that is the current name of the community release of Docker.

Both 'docker version' and 'docker ps -a' output is now collected. Output from
'docker version' help to validate a given docker binary and 'ps -a' output is
collected to see all terminated, but still existing, containers present on the
system. This is kept separate from normal 'ps' output to keep it simple to only
quickly look at running containers.

The output of 'docker inspect' is now collected as well, by default only for
running containers.

Further, plugin options have been changed:

- 'logs' will now capture 'docker logs' output for all running containers.
Previously this was provided by the 'all' option which would also always include
terminated containers.

- 'all' will now cause the plugin to also include terminated containers for
'inspect' and 'logs' output. Note that by itself this option no longer captures any
additional information.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>