[utilities] Support glob expansion in command arguments
Historically, when calling an external process, sos relied on the
use of a shell to perform glob expansion on arguments passed to
the command.

This potentially exposes sos and its children to the entire gamut
of shell expansion and all the problems associated with it and
this is part of the reason that use of a shell in callouts was
disabled in commit 46b6c3d (Issue #253).

While full shell syntax is undesirable for many reasons the use
of globs to express path arguments to external commands is very
useful and widely applied: restore this limited support for
shell-like syntax using the python glob.glob module and expand
any argument that is a valid glob into its resulting arguments.

Fixes: #752

Signed-off-by: Shane Bradley <sbradley@redhat.com>