[rpm] make rpm.rpmva work without --verify
When the global verify switch was introduced the behaviour of the
historical 'rpm.rpmva' plugin option was made to depend on the
global setting: without --verify no verification is carried out
even though it is clear this is not what the user intended.

Allow verification to take place in the rpm module if the rpm.va
option is True, regardless of the setting of the global --verify
option.

Fixes #735.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>