[npm] find modules in npm cache
npm installs modules:

 1. globally — so everyone can use those
 2. per project

If you don't know the path to the project, the only way to see the list
of those modules is by traversing the whole filesystem, which is pretty
unreal.

This patch browses npm cache of every user on the system and saves
module names and all available versions into sosreport.

Signed-off-by: Tomas Tomecek <ttomecek@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>

Closes: #879.