[openstack_*] Allow some log files to be excluded
OpenStack deployments can create machine readable json log files, eg:

/var/log/glance/glance-api.log
/var/log/glance/glance-api-json.log

The json logs contain information which is a duplicate of the equivalent
non-json logs. Both can be quite large.

Add an 'exclude' option to the openstack plugins which allows log files
matching a specification (eg "/var/log/glance/*json*") to be skipped.

This greatly reduces the resulting sos tar file size, the time to gather
the logs, and the amount of space in /tmp required when generating the
tar file.