[sosreport] Change text format, add username, filesize and chksum
This changes the format of the end text making the output information
more consistent.  We now mention the username that owns the archive
the filesize in a human readable format, and the checksum.

New format:
Your sosreport has been generated and saved in:
  /tmp/sosreport-desktop-20180313115538.tar.xz

 Size	63.55KB
 Owner	root
 sha256	d0ae51b7e57a6f6f8b336cc0d5a9b5bcc3fd1f6796cfc0df2e74bb98590b82dc

We remove the tmpdir from the pretext.  We also enable saying if
configuration changes should be expected or not, given options
selected.

Closes: #631

Signed-off-by: Bryan Quigley <bryan.quigley@canonical.com>