[docker] Capture cert, network, sizing and stats data
Captures a directory listing of /etc/docker to show the cert structure for the
daemon.

Now captures network listing and inspect details for configured docker networks.

Adds a new 'size' option that when enabled captures 'ps -as' output. This is
disabled by default as this could be very slow for hosts with large containers.

Now captures output from 'docker stats' for containers running when sos is run.

Closes: #901.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>