[virsh] Collect host, network, nwfilter, pool and more domain info
This patch adds collection of host information such as memory and capabilities.

Additionally, user-defined elements such as networks, nwfilters and storage
pools are now collected. A listing of all elements is taken and then, if they
exist, the element's xml is collected as well.

Further, more domain (vm) information is collected - such as dominfo and
domblklist for each domain reported by virsh.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>