[archive] accept an open File for add_file()
Allow subclasses of FileCacheArchive to accept either a path to
a file to add, or an instance of an open File object, from whih
data will be read.

This will be used to covert logging FileHandler object to
StreamHandler objects, avoiding creating bogus '<fdopen>' paths
in the file system when the former are used with files that are
created with os.fdopen().

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>