[kernel] add /sys/fs/pstore collection for all platforms
The pstore data should be collected upon sosreport run for all
architecture (currently only enabled in powerpc.py) because the
data there will include any past oops event and/or panic events.
This data would assist diagnosing panics with little other data
available.

Also, the data included is a few kb in size, so it will not add
much size to the output of the report.

I just moved the /sys/fs/pstore collection from powerpc.py
to kernel.py in order to have it collect for all architectures.

Signed-off-by: Chris Newcomer <chris@thenewcomers.org>