[policies/redhat] make missing 'filesystem' package non-fatal
On Red Hat systems, we use the version of the installed filesystem'
package to enable handling of UsrMove[1] in the PATH used when
running commands from sos.

If the package is not present (e.g. because of a broken system
configuration, rpm timeout, corrupt database or other cause),
previous versions would fail immediately. In this situation
there was no way for the user to attempt a complete run (short
of fixing the problem related to the 'filesystem' package).

Now that the timeout has been lengthened for this command, make
the UsrMove determination a bit more robust to prevent users from
being unable to run sos at all:

  - If RPM fails entirely, exit (plugin enablement tests will fail)
  - If RPM succeeds but 'filesystem' is missing assume UsrMove
    (this is true for all supported Red Hat distros today)
  - Otherwise apply the normal PATH logic.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>