[block] Add option to get data from all block dev.
A command line called 'all' was defined to this plugin.
This option aims to get the data from all block devices in /sys/block/.

The block plugin will continue to work just as before. The
only difference is that now the user has the option to get info from all
devices.

Signed-off-by: Rodrigo R. Galvão <rodrigorosattig@gmail.com>