[landscape] add /sys/class/dmi/id/sys_vendor
The get_vm_info and _get_vm_by_vendor functions in the Landscape
client's vm_info.py script rely on /sys/class/dmi/id/sys_vendor
for VM type detection.

Sometimes Landscape doesn't detect the VM type, because of an
unknown string in that file. Adding this file will help us
understand why Landscape didn't detect the computer as a VM.

The data included in this file is only a few characters long and
will not have an impact on the time or size of the report.

Signed-off-by: David Coronel <david.coronel@canonical.com>