[Plugin] add executable command enablement checks
Add a new list/tuple member to the Plugin class that contains a
list of executable commands that will enable the plugin if any
are present.

For example, a plugin:

  class MyPlugin(Plugin, RedHatPlugin):

      commands = ('mycmd1', 'mycmd2')

Will be automatically enabled if either 'mycmd1' or 'mycmd2'
is present and executable in the policy defined PATH for the
run.

Related: #1051.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>