[policies] stop execution on Ctrl+C during user input
When pressing Ctrl+C during user input (name and case id), sosreport
treats that as an empty string and starts to collect data. It must
terminate instead.

Catch the KeyboardIntterupt, print a newline (to avoid munging the
terminal too much), and then re-raise the exception to cause the
process to exit.

Allow all other exceptions to propagate up and do not attempt to
continue.

Resolves: #1094

Signed-off-by: Pavel Moravec <pmoravec@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>