[azure] Added instance metadata and curl to RHUI if on demand image.
Instance metadat is useful to investigate the virtual machine
properties. We will find the base image, size, location and so on from
it. Also, connectivity to RHUI repo is required to work property for
Azure RHEL On-Demand image.

We fixed the WALinuxAgent package name type to enable this plugin
by default when WALinuxAgent is running. It's usually running only on
Azure.

Resolves: #1145

Signed-off-by: Takayoshi Tanaka <tatanaka@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>