[postgresql] use a single name for the db dump file
Currently, there are two names for the file that contain
the postgresql db dump. The original one, which is sos_pgdump.tar
and the new one sos_scl_pgdump.tar based on postgresql95.
However, changing the name of file that contain the pgdump
because of new technologies break the tools that are
based on sos tool. This patch make sure we use the same name.

Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>