[sosreport] Fix debian policy PackageManager misusage.
Python does not impose any kind of ordering on keyword args,
but if the keywords are missing it will still attempt to honour
the call by mapping positional call arguments to corresponding
keywords in the declaration.

In this case the Debian policy invokes the PackageManager call
as though it only uses positional arguments. This is undoubtedly
wrong.

Signed-off-by: Eric Desrochers <eric.desrochers@canonical.com>