[sosreport] Concurrently run plugins
Changes sos to run plugins concurrently. By default sos will now use
four (4) threads to run plugins, allowing for faster overall execution
of sosreport. The number of threads can be changed using the --threads
commandline option.

Plugins now also have a timeout applied to them as a whole to avoid
situations where sosreport appears to be hung. If a plugin exceeds the
timeout threshold, the plugin will be terminated immediately. - this
allows sos to not only continue running normally, but should still allow
for collection of commands run by the plugin up until it was terminated.
The timeout is plugin controlled, and defaults to 300 seconds if not
set.

Note that for python2 environments, this adds a dependency on
python-futures. The futures module is present in the standard library
for python3 environments.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>