[sosreport] Allow policies to change report name generation
This changes the way sos generates the name for the final archive.

Policies may now control the naming pattern by changing the new
'name_pattern' attribute. By default, this is set to 'legacy' for all
policies, and maintains the naming pattern that sos has used up until
this point. A policy may also set name_pattern to 'friendly' which will
result in a more user-friendly name for the archive. Additionally, a
policy may set name_pattern to a format() parsable string to explicitly
set how the tarball should be named.

Additionally, this removes the prompt for a 'customer name' on the
report when running interactively. Instead, the short hostname of the
system is now always used in place of the 'customer name'.

A new option, --label, has been added (and which --name will now map to)
which if provided will append a user-given string to the archive name
after the hostname.

This also sets the Red Hat policy to use the new 'friendly' pattern.

Resolves: #469

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>